UP BlackBox 2.0/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
pandas
requests
python-multipart 